requirements.txt
python-dotenv==1.0.0
Telethon==1.33.1
loguru==0.7.2
google-api-python-client==2.38.0
oauth2client==4.1.3